# backend/src/main/java/com/jll/cibus/payment/dto/DiscountRequestDTO.java
package com.jll.cibus.payment.dto;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import lombok.*;

import java.math.BigDecimal;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class DiscountRequestDTO {

    @NotNull
    @Positive
    private BigDecimal amount;

}