# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: CIBUS API Sistema de Gestión de Servicios Gastronómicos
  version: 1.0.0
  description: >
    Documentación de la API REST para el sistema de gestión de servicios gastronómicos CIBUS.
servers:
  - url: http://localhost:8080/api
    description: Servidor local de desarrollo
security:
  - bearerAuth: []
...
paths:
...
  /auth:
    post:
      tags:
        - Auth
      summary: Authenticate user
      description: Authenticates a user and returns an access token and a refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
        '404':
          description: User not found
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Generates a new access token using a valid refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
        '404':
          description: User not found

  # ── BRANCHES ──────────────────────────────────────────────────────────────
  /branches:
    get:
      tags:
        - Branches
      summary: Get all branches
      description: Returns a list of all branches.
      responses:
        '200':
          description: Branches retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BranchResponseDTO'

    post:
      tags:
        - Branches
      summary: Create a branch
      description: Creates a new branch.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchRequestDTO'
      responses:
        '201':
          description: Branch created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponseDTO'
  /branches/{branchId}:
    get:
      tags:
        - Branches
      summary: Get branch by ID
      description: Returns a branch by its ID.
      parameters:
        - $ref: '#/components/parameters/branchId'
      responses:
        '200':
          description: Branch found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponseDTO'

    put:
      tags:
        - Branches
      summary: Update a branch
      description: Updates an existing branch.
      parameters:
        - $ref: '#/components/parameters/branchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchRequestDTO'
      responses:
        '200':
          description: Branch updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponseDTO'

    delete:
      tags:
        - Branches
      summary: Delete a branch
      description: Deletes a branch by its ID.
      parameters:
        - $ref: '#/components/parameters/branchId'
      responses:
        '204':
          description: Branch deleted successfully
  /branches/search:
    get:
      tags:
        - Branches
      summary: Search branch by address
      description: Returns a branch matching the provided street and number.
      parameters:
        - name: street
          in: query
          required: true
          schema:
            type: string
        - name: number
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Branch found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponseDTO'

   # ── PRODUCTS ──────────────────────────────────────────────────────────────
...
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDTO'

    put:
      tags:
        - Products
      summary: Update product
      parameters:
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDTO'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDTO'

    delete:
      tags:
        - Products
      summary: Delete product
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '204':
          description: Product deleted successfully
...
  /branches/{branchId}/products:
    get:
      tags:
        - Branch Products
      summary: Get branch products
      description: Returns a paginated list of products available in a branch.
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: productId
          in: query
          schema:
            type: integer
        - name: productName
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Products retrieved successfully

    post:
      tags:
        - Branch Products
      summary: Add product to branch
      description: Associates a product with a branch.
      parameters:
        - $ref: '#/components/parameters/branchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchProductRequestDTO'
      responses:
        '201':
          description: Product added successfully
  /branches/{branchId}/products/{productId}:
    get:
      tags:
        - Branch Products
      summary: Get branch product
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/productId'
      responses:
        '200':
          description: Product found

    put:
      tags:
        - Branch Products
      summary: Update branch product
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchProductUpdateDTO'
      responses:
        '200':
          description: Product updated successfully

    delete:
      tags:
        - Branch Products
      summary: Delete branch product
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/productId'
      responses:
        '204':
          description: Product removed successfully

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /branches/{branchId}/orders:
    get:
      tags:
        - Orders
      summary: Get orders
      description: Returns a paginated list of orders for a branch.
      parameters:
        - $ref: '#/components/parameters/branchId'
        - name: tableNumber
          in: query
          schema:
            type: integer
        - name: waiterId
          in: query
          schema:
            type: integer
        - name: statusName
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: minTotal
          in: query
          schema:
            type: number
        - name: maxTotal
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Orders retrieved successfully

    post:
      tags:
        - Orders
      summary: Create order
      parameters:
        - $ref: '#/components/parameters/branchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequestDTO'
      responses:
        '201':
          description: Order created successfully
  /branches/{branchId}/orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Order found

    put:
      tags:
        - Orders
      summary: Update order
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/orderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateDTO'
      responses:
        '200':
          description: Order updated successfully
  /branches/{branchId}/orders/{orderId}/payments:
    post:
      tags:
        - Orders
      summary: Add payment to order
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/orderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentDTO'
      responses:
        '201':
          description: Payment added successfully
  /branches/{branchId}/orders/{orderId}/discount:
    patch:
      tags:
        - Orders
      summary: Apply discount to order
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/orderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountRequestDTO'
      responses:
        '200':
          description: Discount applied successfully
  /branches/{branchId}/orders/statuses:
    get:
      tags:
        - Orders
      summary: Get order statuses
      parameters:
      - $ref: '#/components/parameters/branchId'
      responses:
        '200':
          description: Status list retrieved successfully
  /branches/{branchId}/orders/{orderId}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel order
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Order cancelled successfully
...
  /branches/{branchId}/orders/{orderId}/serve:
    post:
      tags:
        - Orders
      summary: Mark order as served
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/orderId'
      responses:
        '200':
          description: Order marked as served

  # ── ORDER DETAILS ─────────────────────────────────────────────────────────
...
components:
  parameters:
    branchId:
      name: branchId
      in: path
      required: true
      description: ID de la sucursal
      schema:
        type: integer
        example: 1
...
    productId:
      name: productId
      in: path
      required: true
      description: ID del producto
      schema:
        type: integer
        example: 10
...
    orderId:
      name: orderId
      in: path
      required: true
      description: ID del pedido
      schema:
        type: integer
        example: 7
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    
  schemas:
...
    AuthRequest:
      type: object
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: admin123
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        AccessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
      required:
        - refreshToken

    # ── BRANCH ──────────────────────────────────────────────────────────────
    BranchRequestDTO:
      type: object
      properties:
        name:
          type: string
          example: "Sucursal Centro"
        street:
          type: string
          example: "San Martín"
        number:
          type: integer
          example: 1234
      required:
        - name
        - street
        - number
    BranchResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Sucursal Centro"
        street:
          type: string
          example: "San Martín"
        number:
          type: integer
          example: 1234

    # ── PRODUCT ─────────────────────────────────────────────────────────────
    ProductRequestDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: "Milanesa Napolitana"
        description:
          type: string
          maxLength: 255
          example: "Breaded beef with tomato sauce and mozzarella cheese"
        categoryId:
          type: integer
          format: int64
          example: 2
      required:
        - name
        - categoryId
    ProductResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Milanesa Napolitana"
        description:
          type: string
          example: "Breaded beef with tomato sauce and mozzarella cheese"
        categoryName:
          type: string
          example: "Main Course"

  # ── PRODUCT CATEGORY ──────────────────────────────────────────────────────────────
...
    BranchProductRequestDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 5
        price:
          type: number
          format: decimal
          example: 2500.00
      required:
        - productId
        - price
    BranchProductUpdateDTO:
      type: object
      properties:
        price:
          type: number
          format: decimal
          example: 2800.00
        available:
          type: boolean
          example: true
...
    OrderRequestDTO:
      type: object
      properties:
        tableNumber:
          type: integer
          example: 12
        waiterId:
          type: integer
          format: int64
          example: 5
      required:
        - tableNumber
        - waiterId
    OrderUpdateDTO:
      type: object
      properties:
        tableNumber:
          type: integer
          example: 15
...
    PaymentDTO:
      type: object
      properties:
        paymentMethodId:
          type: integer
          format: int64
          example: 2
        amount:
          type: number
          format: double
          example: 5000.00
      required:
        - paymentMethodId
        - amount
    DiscountRequestDTO:
      type: object
      properties:
        amount:
          type: number
          format: double
          example: 1500.00
      required:
        - amount

    # ── ORDER DETAIL ─────────────────────────────────────────────────────────
...